Adding Nuclear Meltdown when tile is destroyed
The nuclear powerplant now has a meltdown when destroyed. A specific tile within the powerplant is required to be hit for this to occur. In the future I would like to expand this to all tiles within the power plant. There are also some added comments to aid in my personal understanding of variables.