Fixed RadMonster bug and made monster head directly to destination if not rampaging
Worked on during office hours. Fixed the  bug where RadMonster would not appear as its location was outside the bounds of the map. Edited existing code so that the original monster would head directly to the nuclear tile if there was one, and would rampage normally if not. Also added and finished comments for readability.