Worked on debugging
Started debugging the monster sprite, which flickers/spins when it is moving southwest towards the nuclear powerplant most of the time. I believe there is also an issue with destX and destY for normal monsters as well, as they almost always move towards the upper-left corner. Also cleaned up my print statements to make debugging easier.